moved tests to tests folders and added correct imports for student and tutor and auth service
