+++ backend/src/modules/user-management/auth.service.js
@@ -3,7 +3,7 @@ import crypto from "crypto";
 import bcrypt from "bcrypt";
 import { Op } from "sequelize";
 import { User, RefreshToken } from "../../models/index.js";
-import Agency from "../../models/agency.model.js";
+import Agency from "../../models/index.js";
 
 export default class AuthService {
   constructor() {
